Completed section "RTL Design Rationale".
The section has two illustrations made in yEd
(http://www.yworks.com/en/products_yed_about.html), and I've added
exports in the PDF format so that yEd won't be required to generate
the report.

However, only 'pdflatex' uses PDF for graphics, and document
generation might fail if generating DVI, for example. Exports in the
appropriate formats can be made with yEd to fix such problems, but
I've not added them to the repository because they are large, and
adding more binary derivatives than absolutely necessary is bad
practice when using git.